go to interactor instead useCases
