+++ app/src/main/java/com/example/playlistmaker/search/domain/TracksInteractorImpl.kt
@@ -1,5 +1,9 @@
 package com.example.playlistmaker.search.domain
 
+import android.content.Context
+import android.content.Intent
+import androidx.core.content.ContextCompat.startActivity
+import com.example.playlistmaker.R
 import java.util.concurrent.Executors
 
 class TracksInteractorImpl(
@@ -16,5 +20,13 @@ class TracksInteractorImpl(
                 response = fromRep.responseCode
             )
         }
+    }
+
+    override fun shareClick(context: Context) {
+        val intent = Intent(Intent.ACTION_SEND)
+        intent.setType("text/plain")
+        intent.putExtra(Intent.EXTRA_TEXT, context.getString(R.string.link_to_android_course))
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+        startActivity(context, intent, null)
     }
 }
