+++ app/src/main/java/com/example/playlistmaker/settings/ui/SettingsViewModel.kt
@@ -1,18 +1,21 @@
 package com.example.playlistmaker.settings.ui
 
+import android.content.Context
 import androidx.lifecycle.ViewModel
+import com.example.playlistmaker.search.domain.TracksInteractor
 import com.example.playlistmaker.settings.domain.OpenLinkUseCase
-import com.example.playlistmaker.settings.domain.ShareUseCase
 
 class SettingsViewModel(
     private val openLinkUseCase: OpenLinkUseCase,
-    private val shareUseCase: ShareUseCase
+    private val interactor: TracksInteractor,
+    private val context: Context
 ) : ViewModel() {
 
     fun onSupportClick() {
         openLinkUseCase.execute()
     }
     fun onShareClick() {
-        shareUseCase.execute()
+//        shareUseCase.execute()
+        interactor.shareClick(context)
     }
 }
